Ajuste nas classes de domínio

+++ src/main/java/com/tweber/helpdesk/domain/Chamado.java
@@ -15,7 +15,7 @@ import java.time.LocalDate;
 @Entity
 public class Chamado implements Serializable {
 
-    public Chamado(Integer id, Prioridade prioridade, Status status, String titulo, String observacoes, Tecnico tecnico, Cliente cliente) {
+    public Chamado(Long id, Prioridade prioridade, Status status, String titulo, String observacoes, Tecnico tecnico, Cliente cliente) {
         this.id = id;
         this.prioridade = prioridade;
         this.status = status;
@@ -27,7 +27,7 @@ public class Chamado implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @JsonFormat(pattern = "dd/MM/yyyy")
     private LocalDate dataAbertura = LocalDate.now();
