+++ src/main/java/com/tweber/helpdesk/repositories/ChamadoRepository.java
@@ -3,5 +3,5 @@ package com.tweber.helpdesk.repositories;
 import com.tweber.helpdesk.domain.Chamado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChamadoRepository extends JpaRepository<Chamado, Integer> {
+public interface ChamadoRepository extends JpaRepository<Chamado, Long> {
 }
