+++ src/main/java/com/tweber/helpdesk/repositories/TecnicoRepository.java
@@ -3,5 +3,5 @@ package com.tweber.helpdesk.repositories;
 import com.tweber.helpdesk.domain.Tecnico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TecnicoRepository extends JpaRepository<Tecnico, Integer> {
+public interface TecnicoRepository extends JpaRepository<Tecnico, Long> {
 }
